refactor: 💡 move type about method
