--- lib/masking/config/target_columns/method/date.rb
@@ -1,28 +1,0 @@
-# frozen_string_literal: true
-
-require 'masking/config/target_columns/method/methodable'
-require 'date'
-
-module Masking
-  class Config
-    class TargetColumns
-      class Method
-        class Date
-          include Methodable
-
-          def call(_sql_value)
-            "'#{date_format}'"
-          end
-
-          private
-
-          FORMAT = '%Y-%m-%d'
-
-          def date_format
-            value.strftime(FORMAT)
-          end
-        end
-      end
-    end
-  end
-end
