--- lib/masking/config/target_columns/method/float.rb
@@ -1,19 +1,0 @@
-# frozen_string_literal: true
-
-require 'masking/config/target_columns/method/methodable'
-
-module Masking
-  class Config
-    class TargetColumns
-      class Method
-        class Float
-          include Methodable
-
-          def call(_sql_value)
-            value.to_s
-          end
-        end
-      end
-    end
-  end
-end
