+++ lib/masking/config/target_columns/method/type/null.rb
@@ -4,11 +4,13 @@ module Masking
   class Config
     class TargetColumns
       class Method
-        class Null
-          def initialize(*); end
+        module Type
+          class Null
+            def initialize(*); end
 
-          def call(_sql_value)
-            'NULL'
+            def call(_sql_value)
+              'NULL'
+            end
           end
         end
       end
