+++ spec/masking/config/target_columns/method/type/binary_spec.rb
@@ -2,9 +2,9 @@
 
 require 'spec_helper'
 
-require 'masking/config/target_columns/method/binary'
+require 'masking/config/target_columns/method/type/binary'
 
-RSpec.describe Masking::Config::TargetColumns::Method::Binary do
+RSpec.describe Masking::Config::TargetColumns::Method::Type::Binary do
   describe '#call' do
     subject { described_class.new(value).call('sql_value') }
