+++ spec/masking/config/target_columns/method/type/float_spec.rb
@@ -2,9 +2,9 @@
 
 require 'spec_helper'
 
-require 'masking/config/target_columns/method/float'
+require 'masking/config/target_columns/method/type/float'
 
-RSpec.describe Masking::Config::TargetColumns::Method::Float do
+RSpec.describe Masking::Config::TargetColumns::Method::Type::Float do
   describe '#call' do
     subject { described_class.new(value).call('sql_value') }
