+++ spec/masking/config/target_columns/method/type/string_spec.rb
@@ -2,9 +2,9 @@
 
 require 'spec_helper'
 
-require 'masking/config/target_columns/method/string'
+require 'masking/config/target_columns/method/type/string'
 
-RSpec.describe Masking::Config::TargetColumns::Method::String do
+RSpec.describe Masking::Config::TargetColumns::Method::Type::String do
   describe '#call' do
     subject { described_class.new(value).call('sql_value') }
